Tests fix
