+++ apps/chat-e2e/src/assertions/accountSettingsAssertion.ts
@@ -1,20 +1,11 @@
-import { ElementCaretState, ExpectedMessages } from '@/src/testData';
+import { BaseAssertion } from '@/src/assertions/base/baseAssertion';
 import { AccountSettings } from '@/src/ui/webElements';
-import { expect } from '@playwright/test';
 
-export class AccountSettingsAssertion {
+export class AccountSettingsAssertion extends BaseAssertion {
   readonly accountSettings: AccountSettings;
 
   constructor(accountSettings: AccountSettings) {
+    super();
     this.accountSettings = accountSettings;
-  }
-
-  public async assertCaretState(expectedState: ElementCaretState) {
-    const caret = this.accountSettings.accountSettingsCaret.getElementLocator();
-    expectedState === 'expanded'
-      ? await expect.soft(caret, ExpectedMessages.caretIsExpanded).toBeVisible()
-      : await expect
-          .soft(caret, ExpectedMessages.caretIsCollapsed)
-          .toBeHidden();
   }
 }
