+++ apps/chat-e2e/src/ui/webElements/navigationPanel.ts
@@ -1,4 +1,5 @@
 import { NavigationPanelSelectors } from '@/src/ui/selectors';
+import { AccountSettings } from '@/src/ui/webElements/accountSettings';
 import { BaseElement } from '@/src/ui/webElements/baseElement';
 import { Locator, Page } from '@playwright/test';
 
@@ -6,6 +7,8 @@ export class NavigationPanel extends BaseElement {
   constructor(page: Page, parentLocator: Locator) {
     super(page, NavigationPanelSelectors.container, parentLocator);
   }
+
+  private accountSettings!: AccountSettings;
 
   public backToChatButton = this.getChildElementBySelector(
     NavigationPanelSelectors.backToChatButton,
@@ -16,4 +19,11 @@ export class NavigationPanel extends BaseElement {
   public myWorkspaceButton = this.getChildElementBySelector(
     NavigationPanelSelectors.myWorkspaceButton,
   );
+
+  public getAccountSettings() {
+    if (!this.accountSettings) {
+      this.accountSettings = new AccountSettings(this.page, this.rootLocator);
+    }
+    return this.accountSettings;
+  }
 }
